+++ lib/ui/polls/CreateStadiumPollPanel.dart
@@ -192,7 +192,7 @@ class _CreateStadiumPollPanelState extends State<CreateStadiumPollPanel> {
       child: Container(
           child: PollOptionView(
             title: Localization().getStringEx("panel.create_stadium_poll.text.question", "QUESTION"),
-            hint: Localization().getStringEx("panel.create_stadium_poll.hint.question", "Ask people near you…"),
+            hint: Localization().getStringEx("panel.create_stadium_poll.hint.question", "Ask people…"),
             textController: _questionController,
             maxLength: 120,
             minLines: 3,